Move useOutsideAlerter to its own file
