+++ src/app/components/helperComponents/useOutsideAlerter.jsx
@@ -1,0 +1,15 @@
+import { useEffect } from "react";
+
+export default function useOutsideAlerter(callback, ref) {
+  useEffect(() => {
+    function handleClickOutside(event) {
+      if (ref.current && !ref.current.contains(event.target)) {
+        callback();
+      }
+    }
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, [callback, ref]);
+}
